fix: set database keys to fake values
The database section is not used, but the base chart still tries to set
some sane default values, so to avoid confusion, just override those to
a string value that makes it obvious this section is not used.

The recommended storage location for meters and events is Gnocchi, which
is automatically discovered and used by means of keystone.