fix(gwapi): allow custom nginx.conf
There is no easy way to add `proxy_buffer_size` and similar parameters
to the nginx confs. Though a feature for the same is being worked on and
should be released early November.

This includes the default nginx configuration as deployed in the current
version of nginx gateway fabric, and adds two lines to increase
`proxy_buffer_size` and `proxy_buffers`.

Ref: nginxinc/nginx-gateway-fabric/#1398